Replace TestChecker with plain old Checker
In core Gerrit, we use test types like TestAccount because the
corresponding core types are, for historical reasons, mutable and
otherwise problematic. In the checker APIs, we've been using nice
AutoValue classes from the start. TestChecker contains exactly the same
set of fields as Checker, and we don't envision that changing, so just
get rid of the extra boilerplate. This reduces the work required to add
a new field to Checker.

Leave TestCheckerUpdate, since that contains methods tailored to what is
needed for tests, which is not the same as what's needed elsewhere.

Change-Id: I24b012c2dacbee8556be3a9597a1697bee494661