Add test coverage for NoteDbCheckers#checkersOf
It turns out this method is not yet called from non-test code.
CheckerOperationsImpl intentionally has its own implementation of
checkersOf, so that the tests in e.g. UpdateCheckerIT are depending only
on the test impl, not on the implementation of some core API code
outside of UpdateChecker.

The rest of NoteDbCheckers already has test coverage, because its other
methods are called from REST API handlers. checkersOf fell through the
cracks because we aren't exposing it directly as an API; we might expose
it via some other downstream API, but those aren't implemented yet.

Change-Id: Id547e6367726396ae33e6850580a81a4b7735b65