ListChecks: Include implicit NOT_STARTED relevant checkers
Checkers that are relevant for the change should have their checks
reported to users even if the external system has not yet reported any
check status. Keep track of which checkers associated with the repo are
not found in the check map, and backfill the results with synthetic
entries for any relevant ones.

This backfilling logic only takes effect in the API layer, not the
underlying NoteDb layer. The NoteDb layer is about reporting exactly
what is being stored, and shouldn't need to have a dependency on the
whole change query subsystem to do so.

This change only backfills entries when listing checkers. Backfilling
individual checks when looking up by key will be implemented in a
followup.

Change-Id: I2d105efecb83c0c6c29fa0c8a6eee323d39bdb2f