UpdateCheckIT: Drop tests that don't test updating checks
These tests look like they are testing the code that update checks but
in reality the exception that is tested is thrown by the
Checks#id(CheckerUuid) method and the CheckApi#update(CheckInput) call
is never invoked.

Remove these tests since they give the wrong illusion that the update
checks code is tested.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ie280a7fdab18788743678d8f446538266da1dc35