Do not require Administrate Checkers to view checks on a change
ChecksImpl#id(CheckerUuid) used the checkers API to verify that a
checker exists. This is bad because the checkers API requires the
Administrate Checkers capability to view checkers. Consequently viewing
checks of a change required this capability too.

Requiring Administrate Checkers for viewing checks on a visible change
was not intended. To fix this use the internal API to verify the checker
existence instead.

This also improves the error message when asking for checks of a
non-existing checker.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I883bf7d439e294ecd5d4e00d85a79b000d5b717e