Allow to update check properties without ever setting a state
The API allows to do partial updates on checks and hence the check state
doesn't need to be specified. As a consequence, a check system might
update a check on a change without ever setting the state to e.g.
running as is typically expected. This becomes relevant if the check
system wants to provide further details (e.g. via a future description)
to users before it even has time to run/schedule the check.

Instead of throwing an exception in that situation, we now automatically
assume the 'not started' state, which is consistent with using the
'not started' state for backfilled checks.

Change-Id: I1757f62be507d14e381d41e24ca8cf0b0c5db802