CreateCheckerIT: Remove sandboxing for createCheck test
The createCheck test failed due to timestamp mismatches if other tests
were executed before it. It's still unclear why the setup for the test
that is run first consumes one timestamp more than the setup for
follow-up tests, however we can fix the timestamp assertion by asking
for the current timestamp right before the check is created and then use
this timestamp as expected creation timestamp.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If2cdf8354ac3ca00d14feaac4e28546388f87c5c