Don't require AdministrateCheckers capability for listing pending checks
Every user should be able to list pending checks, but the REST endpoint
should only returns checks on changes that are visible to the caller.

This means when we execute the change query of the checker we must
filter out non-visible changes. To do this we must use
ChangeQueryProcessor instead of InternalChangeQuery to execute the
query.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id6a13154798150d29512afb87150eab637b1826d