Test that listing checks with options works if checkers are invalid
When the ListChecksOption.CHECKER option is used for listing checks
additional fields in CheckInfo are populated ('checkerName', 'blocking'
and 'checkerStatus'). Populating these fields is only possible when the
checker has a valid configuration. If the checker configuration is
invalid we expect that listing checks still works and that the
additional fields are not populated for invalid checkers.

Change-Id: I7b76558600a2254732ad4705d5c2a4044367fb0a
Signed-off-by: Edwin Kempin <ekempin@google.com>