Test that getting a check with options works if checker is invalid
When the ListChecksOption.CHECKER option is used for getting a check
additional fields in CheckInfo are populated ('checkerName', 'blocking'
and 'checkerStatus'). Populating these fields is only possible when the
checker has a valid configuration. If the checker configuration is
invalid we expect that getting the check still works and that the
additional fields are not populated.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib9fafbb105361377eecbd6889deac8eaa94e393a