ChecksSubmitRule: use Check#areAllRequiredCheckersPassing instead
The "IN_PROGRESS" combined check state isn't enough for
the submit rule to tell whether a change is submittable
or not. To resolve this, ChecksSubmitRule can simply
check whether all required checkers have passed through
"Checks#areAllRequiredCheckersPassing".

TODO: we need to find a way to initialize the
CombinedCheckStateCache, which was done by the
ChecksSubmitRule before this change.

Change-Id: Ib1926cb4d4c17966d729ea75b1fe6895f0fbd398