Add a message text field to checks for further details
Checkers should be able to show short messages directly on a check (e.g.
short failure reason; further infos).

For now this is just a bare string. In the long-term, we might consider
to add markdown support but we would have to provide the general
infrastructure for it in Gerrit first (e.g. have markdown rendering for
our hosts) and we would need to do this for other parts of Gerrit as
well.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If51d31b79ff5859aa4fd6dceda248ea2397e96f0