AbstractCheckersTest: Make projectOperations private
Tests that need ProjectOperations should directly inject it and not rely
on the field in the parent class. This fixes a couple of warnings in
sub-classes that already got ProjectOperations injected, which occurred
because the sub-classes used the same field name and hence hid the field
from the parent class.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I34405b548d851dc94b815ffdf9b18fb93a33c9ae