Make name of checker mandatory
So far it was possible to create checkers without name, but unsetting a
name of an existing checker was not allowed. Make this consistent and
require a name also on checker creation. The documentation already
claimed that it was mandatory.

The same as for all other mandatory checker properties, a missing name
makes a checker invalid and it cannot be accessed through the API. This
means existing checkers without name have to be fixed manually to work
after this change. This is OK since the checks plugin is not widely used
yet.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I08822edb6d068d750b3905bc02e01d44914dfc94