Send email notification if combined check state is updated
When the combined check state of a patch set changes by
creating/updating a check we now send an email notification to the
change owner, the reviewer, users who starred the change and project
watchers.

We only send a notification when the combined check state changes
because sending a notification each time a check is created or updated
would result in too many emails and we don't want to spam users.

To handle the sending of the email notifications we introduce a new
layer for updating checks that is in between the REST layer and the
storage layer. This way we neither need to reimplement sending email
notifications for all REST callers nor for all storage backends. The
new ChecksUpdate layer delegates the persistence of checks to the checks
storage backend and implements business logic (such a s sending emails)
that is the same for all storage backends. The interface for the
storage layer was renamed from ChecksUpdate to ChecksStorageUpdate (to
free up the ChecksUpdate name for the new layer).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ibb5cd800b786f6754fcf4f3f28f26c2cad7ded94