Add .gitreview file
For folks used to using the git-review tool for interacting with
gerrit servers, it's necessary to have a .gitreview file in the
repo in question pointing to where the gerrit is.

Add one, pointing at gerrit-review.googlesource.com and listing that
https should be used. This will still require that someone add their
http authentication to the .gitcookies file, but otherwise should
work as expected.

A similar file was recently added to Gerrit itself.

Change-Id: I0f537704a41157e5deda16ae103145b896a3d34c