Remove retrying index queries on StorageException
These retries were added to mitigate temporary issues with the index
backend that is used at Google that resulted in throwing
StorageException. This case is now handled in a Google-specific
ExceptionHook implementation so that upstream Gerrit no longer needs to
retry index queries on StorageException. Removing this is good because
StorageException may also be thrown in other cases that should not
trigger a retry.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6ba5280489005e412e103c7d5cdb2221121d1268