Adapt to removal of RetryHelper#indexQuery method
Change I9b577a827 removes the RetryHelper#indexQuery method in favor of
more specific methods to execute account/change queries
(RetryHelper#accountIndexQuery and RetryHelper#changeIndexQuery). The
new methods provide an InternalQuery to execute the index query. For
CheckerQuery the new methods are not suitable since it uses
ChangeQueryProcessor instead of InternalChangeQuery to query the index.
Rewrite the code to use the generic RetryHelper#action method instead.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If7a580bff70483f2c3b97d8f0797373c7616a597