Replace Set<BlockingCondition> with "required" boolean
BlockingCondition was originally implemented as a checker field used to
describe conditions that block a submission. While this is useful for
Checkers, it is not useful for Checks since the frontend needs to
interpret them correctly.

This change creates an alternative field that directly indicates whether
this specific check is relevant for submit of a change. Also, this helps
the backend more accurately reflect the submit impact as considered within
the submit rule.

No storage changes are necessary and hense no mgiration is required.

Follow-up changes will have the "required" field entering another Info
entity for the purpose of having it easier to add more fields which are
in direct relation to whether a change can be submitted, by having all
those fields grouped under the same entity.

Another follow-up change will have documentation updates.

This change only adds an alternative to BlockingCondition without
removing support for it. Thus, it doesn't break the frontend.
However, the frontend should still adjust by using the new field. Once
the frontend is done, I can clean the backend by removing the legacy
BlockingCondition, as it is used only by the frontend.

Change-Id: I1790c62a6ee048fb7171b8a43ca95ca370dda210