CheckSubmitRule: Return Optional#empty() if all checkers are optional
If all configured checkers for a project are optional, it's better to
return an Optional#empty() instead of a record with status=OK.

This is also needed by submit requirements, since we convert results of
legacy submit rules to submit requirements. If we return an
Optional#empty(), the SR evaluator will just skip creating a submit
requirement result for it.

Google-Bug-Id: b/209752314
Change-Id: Ia96861ed4bd5b4644f79988e3a7a8d221d42a227