ChecksUpdate: use ChangeIndexer to index changes
We saw in the logs thrown exceptions on this code path with message:
  * AuthException: maintain server not permitted

This was because the current code delegated to ChangeApiImpl which calls
Index (see [1]), which in turn checks that the calling user has the
'Maintain Server' global capability and fails. This failed the POST
request to update checks.

Updating the code to use ChangeIndexer instead.

[1] https://gerrit.googlesource.com/gerrit/+/88f26343b13560d4c75b25a77a59427972c1c91a/java/com/google/gerrit/server/restapi/change/Index.java#44

Google-Bug-Id: b/250877007
Change-Id: Iea638a383ecfc7ddf6a3b46ccc6a005eb5cd3aea